Organize front, building pages again
