+++ my-app/src/components/Bill.tsx
