+++ my-app/src/components/PaymentMethod.tsx
