+++ my-app/src/index.tsx
@@ -3,7 +3,7 @@ import { ChakraProvider } from '@chakra-ui/react';
 
 import ReactDOM from "react-dom/client";
 import "./styles/index.scss";
-import App from "./App";
+import App from "./components/App";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
