+++ my-app/src/pages/Checking.tsx
@@ -1,0 +1,1 @@
+export {};
