+++ my-app/src/pages/Payment.tsx
@@ -1,0 +1,1 @@
+export {};
