+++ my-app/src/styles/index.scss
@@ -16,8 +16,6 @@ $on-hover-grey : #6D6D6D;
 
 body {
     margin: 0;
-    padding-left: 200px;
-    padding-top: 200px;
     font-family: "Raleway", -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
     "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
     sans-serif;
@@ -35,6 +33,26 @@ code {
     display: flex;
     flex-direction: row;
     gap: 300px;
+}
+
+.login-section{
+    margin: 0px;
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    height: 100%;
+}
+
+.decorative-container{
+    display: inline-flex;
+    height: 100vh;
+    width: 50%;
+    background-color: red;
+
+    img{
+        height: 30%;
+        width: 30%;
+    }
 }
 
 .PaymentMenu{
